Add player class and object
Add player class, of type Actor and InputHandler.

Add actor class, which will be an abstraction for any player - including
artifically intelligent players.

Make a player object in main. Player class uses a Character*, and sends
it to the Actor class.

In main, assume the one player object is the only one handling keyboard
events at the moment, and use its handle procedure to handle events.

Rename ACTIONTYPE_LAST to NUM_ACTIONTYPE, which is a more logical
candidate for a sizeOf.

Change controllers.conf to have sample keycodes in valid range.

Read keybinds in main before running the game.