Create Game class, add third player, & minor fixes
Create Game, and put game loop there instead of in bin file.

Add third player for testing purposes. It is controlled by keyboard -
,aoe. Second player is still gamepad-controlled.

The camera is borked for more than two players. This is highlighted in
the code, and very much visible in the game itself.

Change Init to ConfigParser.

Create dedicated physics and graphics directories.

Create config directory.

Fix the Makefile. It will now put stuff in /usr/share/local/limbs-off/.
The config files will be in /usr/share/local/limbs-off/config/. A data
directory should be created when we have graphics and sounds. #defines
for all of this is created in the Makefile. We should take ~ into
consideration as well in the future.