Separate handling and generation of events in Menu
Push events in Menu::raiseEvent(EVENT_ID id), instead of Menu::handle().

Create an enum for different event IDs (can be an event type, or in the
future a code for a SDL_USEREVENT).

Make Menu::handle() return true if it handled the event, and false if it
did not.

Make the getGuiElement and the getSelected (both in Menu) procedures
private, as there is no reason for them to be public.