Handle collisions between characters
Add collision groups so that characters don't collide with themselves.

Add operators for adding and scaling body states, and functions for
getting and setting the body state in Body.

Store inverse mass in Mass instances and use that for some of the
calculations. Model immovable objects as having an inverse mass of 0.

Add a bounce2 function, which handles collisions between two movable
bodies.

Add Collision and CollisionQueue classes to store collision details and
sort them by time. Handle character-to-character and character-to-planet
collisions every step by using the bounce functions.