Change character mass on impact and in space
Implement a collision handler to sort collisions. Monitor character body
parts for hits with this. Upon collision, change the mass of the main
body part of the character based on the force of the impact.

When the character has lost contact with the planet - reduce its main
body's mass to prevent it becoming stuck, floating in space forever.

Set a minimum mass for the character, and kill it when it reaches this
mass by disabling the limb links.

Adjust the roll resistance to make the smaller limbs roll on the planet.

Make the camera only follow alive characters. Set it to an appropriate
zoom level when all characters are dead.

Implement setRadius to change the radii of circles. Implement
SizeModifier to change the sizes of graphics. Use this to our advantage
by updating the shape and the character's graphic to reflect the change
in mass with visual feedback.

Change the initial values for characters' mass, shapes and graphics
radii to better fit with the change mass mechanic. Change the formula
for crouching to make sure fat characters don't get their ass stuck.