Add suicide action
Add a suicide trigger for each player. Mainly for testing purposes.

Adding this option to the gamepad brought out a latent bug. It seems
that SDL_JoystickOpen() actually simulates button presses for all
gamepad buttons. This means that for the first game that is created when
the executable is run - the gamepad-controlled character will commit
suicide after receiving an event on the suicide-mapped button.

This commit prevents this bug by moving the SDL_JoystickOpen() from Game
to the bin.