Implement mass indicators
Implement MassIndicator and MassIndicatorGraphic. The former holds the
logic, the latter the graphic.

Add the GuiElement and HudElement subclasses of ScreenElement to
properly distinguish the two.

Create Character::getMass(). Use this when updating the correlated mass
indicator text.

Implement horizontal functionality and perpendicular offset for
PositionModifier. Use this to distribute the MassIndicatorGraphics
horizontally at the bottom of the screen.

Implement CharacterGraphic::getColourModifier(). Use CharacterGraphic's
ColorModifier as the correlated MassIndicatorGraphic's ColorModifier.

Free Label's surface so that we don't leak.

Handle OpenGL colours correctly in Label::doDraw().